Improve feedback when selecting condition type
Make selection of condition type and their forms closer together.
