+++ graylog2-web-interface/src/components/event-definitions/event-definition-form/EventDetailsForm.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import lodash from 'lodash';
 import { Col, ControlLabel, FormGroup, HelpBlock, Row } from 'react-bootstrap';
-import { PluginStore } from 'graylog-web-plugin/plugin';
 
 import { Select } from 'components/common';
 import { Input } from 'components/bootstrap';
@@ -20,13 +19,6 @@ class EventDetailsForm extends React.Component {
     onChange: PropTypes.func.isRequired,
   };
 
-  getConditionPlugin = (type) => {
-    if (type === undefined) {
-      return {};
-    }
-    return PluginStore.exports('eventDefinitionTypes').find(edt => edt.type === type) || {};
-  };
-
   handleChange = (event) => {
     const { name } = event.target;
     const { onChange } = this.props;
@@ -36,18 +28,6 @@ class EventDetailsForm extends React.Component {
   handlePriorityChange = (nextPriority) => {
     const { onChange } = this.props;
     onChange('priority', lodash.toNumber(nextPriority));
-  };
-
-  handleEventDefinitionTypeChange = (nextType) => {
-    const { onChange } = this.props;
-    const conditionPlugin = this.getConditionPlugin(nextType);
-    const defaultConfig = conditionPlugin.defaultConfig || {};
-    onChange('config', { ...defaultConfig, type: nextType });
-  };
-
-  formattedEventDefinitionTypes = () => {
-    return PluginStore.exports('eventDefinitionTypes')
-      .map(type => ({ label: type.displayName, value: type.type }));
   };
 
   render() {
@@ -84,17 +64,6 @@ class EventDetailsForm extends React.Component {
                       clearable={false}
                       required />
               <HelpBlock>Choose the priority for Events created from this Definition.</HelpBlock>
-            </FormGroup>
-
-            <FormGroup controlId="event-definition-priority">
-              <ControlLabel>Condition Type</ControlLabel>
-              <Select placeholder="Select a Condition Type"
-                      options={this.formattedEventDefinitionTypes()}
-                      value={eventDefinition.config.type}
-                      onChange={this.handleEventDefinitionTypeChange}
-                      clearable={false}
-                      required />
-              <HelpBlock>Choose the type of Condition the Event will use.</HelpBlock>
             </FormGroup>
           </fieldset>
         </Col>
